Shall be refactored later.
