+++ sdk/src/main/java/com/podio/sdk/domain/LinkedAccountData.java
@@ -1,9 +1,11 @@
-package com.podio.sdk.domain.data;
+package com.podio.sdk.domain;
+
+import com.podio.sdk.domain.data.Data;
 
 /**
  * Created by sai on 8/31/16.
  */
-public class LinkedAccountData {
+public class LinkedAccountData implements Data{
     private final String url = null;
     private final String info = null;
     private final String type = null;
